[REFACT] 코드 정리
