+++ src/main/java/com/programmers/smrtstore/domain/user/presentation/facade/UserFacadeImpl.java
@@ -1,7 +1,7 @@
-package com.programmers.smrtstore.domain.user.presentation.controller;
+package com.programmers.smrtstore.domain.user.presentation.facade;
 
-import com.programmers.smrtstore.domain.user.application.ShippingAddressService;
-import com.programmers.smrtstore.domain.user.application.UserService;
+import com.programmers.smrtstore.domain.user.application.service.ShippingAddressService;
+import com.programmers.smrtstore.domain.user.application.service.UserService;
 import com.programmers.smrtstore.domain.user.presentation.dto.req.DetailShippingRequest;
 import com.programmers.smrtstore.domain.user.presentation.dto.req.UpdateShippingRequest;
 import com.programmers.smrtstore.domain.user.presentation.dto.req.UpdateUserRequest;
